refactor: modularize git operations into feature-specific files
- Split monolithic git/git.go into smaller, focused files:
  - git/config.go: Git configuration operations
  - git/rev-parse.go: Rev-parse related operations
  - git/describe.go: Describe operations
  - git/repository.go: Repository information operations
- Move CheckoutNewBranch implementation to git/branch.go
- Align with existing modular design pattern (git/add.go, git/commit.go)
- Improve code maintainability and single responsibility principle
- All tests continue to pass after refactoring
