+++ git/git.go
@@ -1,10 +1,7 @@
 package git
 
 import (
-	"fmt"
-	"os"
 	"os/exec"
-	"strings"
 )
 
 // Client is a git client.
@@ -129,121 +126,9 @@ func NewClient() *Client {
 	}
 }
 
-// Repository Information
-
-// GetCurrentBranch gets the current branch name.
-func (c *Client) GetCurrentBranch() (string, error) {
-	if c.GetCurrentBranchFunc != nil {
-		return c.GetCurrentBranchFunc()
-	}
-	cmd := c.execCommand("git", "rev-parse", "--abbrev-ref", "HEAD")
-	out, err := cmd.Output()
-	if err != nil {
-		return "", NewError("get current branch", "git rev-parse --abbrev-ref HEAD", err)
-	}
-	branch := strings.TrimSpace(string(out))
-	return branch, nil
-}
-
-// RevParseVerify checks whether the given ref resolves to a valid object.
-// It runs: git rev-parse --verify --quiet <ref>
-func (c *Client) RevParseVerify(ref string) bool {
-	cmd := c.execCommand("git", "rev-parse", "--verify", "--quiet", ref)
-	if err := cmd.Run(); err != nil {
-		return false
-	}
-	return true
-}
-
-// GetBranchName gets branch name.
-func (c *Client) GetBranchName() (string, error) {
-	cmd := c.execCommand("git", "rev-parse", "--abbrev-ref", "HEAD")
-	out, err := cmd.Output()
-	if err != nil {
-		return "", NewError("get branch name", "git rev-parse --abbrev-ref HEAD", err)
-	}
-	return strings.TrimSpace(string(out)), nil
-}
-
-// GetGitStatus gets git status.
-func (c *Client) GetGitStatus() (string, error) {
-	cmd := c.execCommand("git", "status", "--porcelain")
-	out, err := cmd.Output()
-	if err != nil {
-		return "", NewError("get status", "git status --porcelain", err)
-	}
-	return string(out), nil
-}
-
-// Branch Operations
-
-// CheckoutNewBranch creates a new branch and checks it out.
-func (c *Client) CheckoutNewBranch(name string) error {
-	cmd := c.execCommand("git", "checkout", "-b", name)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	if err := cmd.Run(); err != nil {
-		return NewError("checkout new branch", fmt.Sprintf("git checkout -b %s", name), err)
-	}
-	return nil
-}
-
-// Config Operations
-
-// ConfigGet retrieves a git configuration value from local repository
-func (c *Client) ConfigGet(key string) (string, error) {
-	cmd := c.execCommand("git", "config", key)
-	out, err := cmd.Output()
-	if err != nil {
-		return "", NewError("config get", fmt.Sprintf("git config %s", key), err)
-	}
-	return strings.TrimSpace(string(out)), nil
-}
-
-// ConfigSet sets a git configuration value in local repository
-func (c *Client) ConfigSet(key, value string) error {
-	cmd := c.execCommand("git", "config", key, value)
-	if err := cmd.Run(); err != nil {
-		return NewError("config set", fmt.Sprintf("git config %s %s", key, value), err)
-	}
-	return nil
-}
-
-// ConfigGetGlobal retrieves a git configuration value from global config
-func (c *Client) ConfigGetGlobal(key string) (string, error) {
-	cmd := c.execCommand("git", "config", "--global", key)
-	out, err := cmd.Output()
-	if err != nil {
-		return "", NewError("config get global", fmt.Sprintf("git config --global %s", key), err)
-	}
-	return strings.TrimSpace(string(out)), nil
-}
-
-// ConfigSetGlobal sets a git configuration value in global config
-func (c *Client) ConfigSetGlobal(key, value string) error {
-	cmd := c.execCommand("git", "config", "--global", key, value)
-	if err := cmd.Run(); err != nil {
-		return NewError("config set global", fmt.Sprintf("git config --global %s %s", key, value), err)
-	}
-	return nil
-}
-
-// GetVersion gets the git version/tag information
-func (c *Client) GetVersion() (string, error) {
-	cmd := c.execCommand("git", "describe", "--tags", "--always", "--dirty")
-	out, err := cmd.Output()
-	if err != nil {
-		return "dev", nil // Return "dev" as fallback instead of error
-	}
-	return strings.TrimSpace(string(out)), nil
-}
-
-// GetCommitHash gets the short commit hash
-func (c *Client) GetCommitHash() (string, error) {
-	cmd := c.execCommand("git", "rev-parse", "--short", "HEAD")
-	out, err := cmd.Output()
-	if err != nil {
-		return "unknown", nil // Return "unknown" as fallback instead of error
-	}
-	return strings.TrimSpace(string(out)), nil
-}
+// All git operations have been moved to appropriate files:
+// - Repository Information: git/repository.go and git/rev-parse.go
+// - Branch Operations: git/branch.go
+// - Config Operations: git/config.go
+// - Describe Operations: git/describe.go
+// - Rev-parse Operations: git/rev-parse.go
